Discover and map GroupVersionResource from API server fix #11
Correctly determining the GVR from a GVK requires communication
with the API server we're connected to.

This PR passes the regular kubernetes client through, extends
getGVR to dynamically determine the Resource and adds a test to
validate this using an ingress resource. Ingress resources were
the first to surface this bug, but are unlikely the only affected
kind.