Cache `restmapper.GetAPIGroupResources` response in provider
To be able to map a resource kind to a REST url, the go client
requires the list of resources the current cluster supports.

The previous implementation naively makes a request to the K8s
API before every request. This results in 259 calls to
`restmapper.GetAPIGroupResources` even though the list rarely
changes. I'm assuming it only does when a custom resource gets
added or removed, or when the K8s API is updated.

Even with the relatively high QPS and Burst values we set, this
frequently resulted in requests being throttled, slowing the
provider down.

This change reduces the number of calls during tests from 259
to 120 by caching the response in memory.