Fix incorrect resource order due to hash > max uint32
Resource IDs that hashed to large numbers potentially exceeded the
max of uint32 due to a bug in the prefixing always appending, not
replacing characters.

Due to the fallback implemented, these resources then ended up
with a hash without prefix, which if that resource was a namespace
led to the issue the prefixing exists to prevent.

This PR improves the prefixing and adds unit tests to test more
cases then the acceptance tests test for.