Fall back to MergePatch if StrategicMergePatch returns invalid fix #86
We try `StrategicMergePatch` first and are supposed to fall back to `MergePatch` when `StrategicMergePatch`
is not an option. But previously we would only try `MergePatch` if the error returned was explicitly
`UnsupportedMediaType`.

However, if the API returns Invalid for `StrategicMergePatch` we have to try `MergePatch` before returning
an error to the user.