Refuse to import resources without LastAppliedConfigAnnotation fix #85
The provider relies on the same LastAppliedConfigAnnotation as kubectl. This
was a decision I made early, that may very well have been unnecessary and
could be changed going forward.

However, I do not currently have the time to look into what this means,
especially for diff output and existing states.

For now, I make the limitation explicit by returning a descript error
during import, instead of failing with a non-descript JSON parse error.