Refactor OverlayFS to handly nothing but our specific use case fix #92
The initial implementaiton tried to implement an in-memory overlay,
over the files in the actual filesystem using the filesys.FileSystem
interfaces from Kustomize.

However, there is an issue with how upstream handles path seperators
on Windows in combination with our use-case here.

We do not need a full in-memory overlay. We really only need
to fake a single Kustomization file and otherwise let Kustomize
use the regular on-disk filesystem.

This commit refactors the overlayFileSystem to not depend on Kustomize's
in-memory FileSystem implementation anymore and instead write the file
to a unique temp directory.