Determine patchType based on GVK and generate correct patch
This commit replaces the previous implementation, that would
first try a StrategicMergePatch and if that failed fall back
to a MergePatch with an implementation that follows the same
logic as kubectl.

It now determines based on the GVK, if it can create a strategic
merge patch. If not, it generates a merge patch. The patch
type is further now returned alongside the patch. Making sure
we're sending the correct patch with the correct patch type.

The previous implementation, incorrectly, did always send a merge
patch, even if the type was set to StrategicMergePatchType.