Require one of `kubeconfig_path` or `kubeconfig_raw` fix #90
This commit changes the provider to require exactly one of
`kubeconfig_path` or `kubeconfig_raw` set. The environment variable
`KUBECONFIG_PATH` is available to set `kubeconfig_path`.

Previously, the provider tried to find a valid configuration
in either `kubeconfig_raw` or `kubeconfig_path` and would fall
back to `kubectl` default environment variables and paths if
no configuration was given.