Fix flaky tests by removing upstream ingress setup
The issue is related to how the Kubestack framework default ingress
currently is set up. This will need to be adressed upstream, but does
not need to cause the provider tests to be flaky.

The ingress base is now replaced by a test base. This is enough
to validate provider changes do not introduce breaking changes
to how the framework configures the provider.