ForceNew when name or namespace change without state ID change fix #132
This issue is rare, because when using `for_each` the state ID will
always change when the name or namespace change.

But when not using `for_each` Terraform will run the diff func and
it can't rely on the dry-run because that will return errors
when the name and namespace in the URL differ from the manifest.