Fix ID format migration documentation #139
I wasn't able to make a regex work using `sed`, because of
all the escaping necessary between the regex and bash.

So instead, I opted to do this in Python, and made the script
print the `terraform state mv` command, instead of executing it.
This way, users can first inspect the commands and then decide
to run them.