Fix CRD tests
The CRD tests have always added the test-key attribute. But this
attribute was not set in the CRD's spec. And hence, the test-key
attribute has never been persisted by the K8s API.

Because the provider only stores the last config it applied
in the Terraform state, and only diffs against that, this bug
never caused the tests to fail.