Compress lastAppliedConfig if metadata.annotations is too long
If metadata.annotations exceeds the max size K8s allows, like
for e.g. in the case of the Prometheus CRD, the provider will
now gzip and base64 encode the annotation value. It will use
the 'kustomization.kubestack.com/last-applied-config-gzip'
key in that case.

Since the provider will only reconcile if the annotation
changes, it will continue to use the regular
'kubectl.kubernetes.io/last-applied-configuration' key
in all other cases.

This is only meant to buy time until the provider can be
migrated to use server side applies.